pyproject.toml
[build-system]
requires = ["poetry-core>=0.12"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "finops_crawler"
version = "0.1.17"
description = "It's a package for importing cost and usage data from various cloud platforms and SaaS tools"
authors = ["Lauri Koobas <laurikoobas@gmail.com>", "FinOps Fitness Club <finops.fitness.club@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
boto3 = "1.26.156"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^1.0.0"
tomlkit = "^0.11.8"

[tool.poetry.urls]
"Home" = "https://github.com/finops-fitness-club/finops_crawler"
"Repository" = "https://github.com/finops-fitness-club/finops_crawler.git"
"Documentation" = "https://github.com/finops-fitness-club/finops_crawler#readme"
